requirements.txt
a2a-sdk==0.3.3
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.64.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
boto3==1.40.19
botocore==1.40.19
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
distro==1.9.0
fastapi==0.116.1
Flask==3.1.2
frozenlist==1.7.0
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.75
langchain-text-splitters==0.3.9
langsmith==0.4.20
MarkupSafe==3.0.2
multidict==6.6.4
openai==1.102.0
orjson==3.11.3
packaging==25.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sse-starlette==3.0.2
starlette==0.47.3
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
Werkzeug==3.1.3
yarl==1.20.1
zstandard==0.24.0
